+++ app/mixins/jam-document-adapter.js
