+++ app/serializers/account.js
@@ -1,9 +1,10 @@
 import DS from 'ember-data';
 
 import JamSerializer from '../mixins/jam-serializer';
-import JamCollectionSerializer from '../mixins/jam-collection-serializer';
+import JamDocumentSerializer from '../mixins/jam-document-serializer';
 
-export default DS.JSONAPISerializer.extend(JamSerializer, JamCollectionSerializer, {
-  modelName: 'account',
-  relationAttrs: ['sessions'],
+export default DS.JSONAPISerializer.extend(JamSerializer, JamDocumentSerializer, {
+    modelName: 'account',
+    //relationAttrs: ['sessions'],
+    // TODO: Known issue- weird hard to trace ember issues coming from how we track relationships
 });
