+++ app/serializers/admin.js
@@ -2,9 +2,9 @@ import Ember from 'ember';
 import DS from 'ember-data';
 
 import JamSerializer from '../mixins/jam-serializer';
-import JamCollectionSerializer from '../mixins/jam-collection-serializer';
+import JamDocumentSerializer from '../mixins/jam-document-serializer';
 
-export default DS.JSONAPISerializer.extend(JamSerializer, JamCollectionSerializer, {
+export default DS.JSONAPISerializer.extend(JamSerializer, JamDocumentSerializer, {
   modelName: 'admin',
-  relationAttrs: ['experiments'],
+  //relationAttrs: ['experiments'],
 });
