+++ app/serializers/session.js
@@ -1,8 +1,8 @@
 import DS from 'ember-data';
 
 import JamSerializer from '../mixins/jam-serializer';
-import JamCollectionSerializer from '../mixins/jam-collection-serializer';
+import JamDocumentSerializer from '../mixins/jam-document-serializer';
 
-export default DS.JSONAPISerializer.extend(JamSerializer, JamCollectionSerializer, {
+export default DS.JSONAPISerializer.extend(JamSerializer, JamDocumentSerializer, {
     modelName: 'session'
 });
